feat: Add complete business layer with auth, services and repositories
- Implement JWT authentication and Spring Security integration
- Add all controllers, services and repositories
- Add DTO/Mapper pattern with MapStruct
- Add Redis support and file upload functionality
- Add Swagger API documentation
- User entity implements UserDetails interface
- Add project documentation and integration tests

+++ src/main/java/com/fooddel/entity/Order.java
@@ -7,6 +7,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import java.math.BigDecimal; // BigDecimalをインポート
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -31,7 +32,7 @@ public class Order {
     private LocalDateTime orderDate;
 
     @Column(name = "total_amount", nullable = false, precision = 10, scale = 2)
-    private Double totalAmount;
+    private BigDecimal totalAmount;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
